+++ tests/cli.rs
@@ -1,5 +1,5 @@
 // Conserve backup system.
-// Copyright 2016-2024 Martin Pool.
+// Copyright 2016-2026 Martin Pool.
 
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by